Rename model_configs.py to models_configs.py for consistency
